Repository Creation
Lost original LOCAL repository due to a computer reset.
Didn´t publish  any previous versions before this point.

This version has a bugged code which seems to get stuck on start and
trigger the error LED.

The main change is to use a timer based on processor cycles instead of
an actual delay (when playing the stored signal).